Fixed some type mismatches

+++ src/utilities/normal_instructions.rs
@@ -14,48 +14,48 @@ impl crate::utilities::cpu::Cpu{
     ///2NNN instruction implementation
     ///Returns from a subroutine by performing a pop on a stack
     pub fn return_from_subroutine(mut self) -> Result<(), &'static str>{
-        self.pc = self.stack.stack[self.stack.sp-1];
+        self.pc = self.stack.stack[self.stack.sp as usize-1];
         self.stack.pop()?;
         Ok(())
     }
     ///3XNN instruction implementation
     ///Skips one instruction if value in register VX is equal to NN
     pub fn skip_if_equal_to(mut self, x: u8, nn: u8){
-        if self.registers[x] == nn{
+        if self.registers[x as usize] == nn{
             self.pc += 2;
         }
     }
     ///4XNN instruction implementation
     ///Skips one instruction if value in register VX is NOT equal to NNN
     pub fn skip_if_not_equal_to(mut self, x: u8, nn: u8){
-        if self.registers[x] != nn{
+        if self.registers[x as usize] != nn{
             self.pc += 2;
         }
     }
     ///5XY0 instruction implementation
     ///Skips one instruction if value in register VX is equal to value in VY
     pub fn skip_if_equal(mut self, x: u8, y: u8){
-        if self.registers[x] == self.registers[y]{
+        if self.registers[x as usize] == self.registers[y as usize]{
             self.pc += 2;
         }
     }
     ///9XY0 instruction implementation
     ///Skips one instruction if value in register VX is NOT equal to value in VY
     pub fn skip_if_not_equal(mut self, x: u8, y: u8){
-        if self.registers[x] != self.registers[y]{
+        if self.registers[x as usize] != self.registers[y as  usize]{
             self.pc += 2;
         }
     }
     ///6XNN instruction implementation
     ///Sets the register VX to the value of NN
     pub fn set_register(mut self, x: u8, nn: u8){
-        self.registers[x] = nn;
+        self.registers[x as usize] = nn;
     }
     ///7XNN instruction implementation
     ///Adds the value NN to register VX
     ///If attempts to add with overflow, only lower 8 bits will be stored 
     pub fn add(mut self, x: u8, nn: u8){
-        self.registers[x] = (self.registers[x] as u16 + nn as u16).to_be_bytes()[1];
+        self.registers[x as usize] = (self.registers[x as usize] as u16 + nn as u16).to_be_bytes()[1];
     }
     ///ANNN instruction implementation
     ///Sets the index register to the value of NNN
@@ -68,7 +68,7 @@ impl crate::utilities::cpu::Cpu{
     ///Sets VF to 1 if addition overflowed 12 bits
     //TODO - make this behaviour changeable?
     fn add_to_index(mut self, x: u8){
-        self.i += self.registers[x];
+        self.i += self.registers[x as usize] as u16;
         if self.i >= 0x1000{
             self.registers[15] = 1
         }
